+++ tests/test_distributed.cpp
@@ -176,7 +176,7 @@ bool matrices_match(const Matrix &lhs, const Matrix &rhs)
 class LogisticLossFunction final : public LossFunction
 {
 public:
-    std::shared_ptr<math::ASTNode> sample_loss(const Matrix &x, int y) const override
+    std::shared_ptr<math::ASTNode> sample_loss(const Matrix &x, double y) const override
     {
         std::shared_ptr<math::ASTNode> logits = math::Num(0.0);
         for (std::size_t i = 0; i < x.numCols(); ++i)
@@ -186,8 +186,8 @@ public:
 
         logits = logits + math::Var("b");
         std::shared_ptr<math::ASTNode> y_hat = math::Sigmoid(logits);
-        return -math::Num(static_cast<double>(y)) * math::Log(y_hat) -
-               (math::Num(1.0) - math::Num(static_cast<double>(y))) * math::Log(math::Num(1.0) - y_hat);
+        return -math::Num(y) * math::Log(y_hat) -
+               (math::Num(1.0) - math::Num(y)) * math::Log(math::Num(1.0) - y_hat);
     }
 
     std::shared_ptr<math::ASTNode> regularizer() const override
